feat: allow to encode a registry in openmetrics without EOF (#27)
## Description

`MetricsSource::encode_openmetrics` adds the `# EOF\n` marker at the end
of the encoded openmetrics output. This is the expected last characters
in the openmetrics text format. Currently,
`MetricsSource::encode_openmetrics` is the only public API to encode a
`iroh_metrics::Registry` into the openmetrics text format.

When combining output from different sources, users should be able to
add the EOF marker themselves. This PR adds a public function
`Registry::encode_openmetrics_to_writer` that encodes without the EOF.

For 1.0 we should revisit the MetricsSource trait, but the change in
this PR is semver compatible so can be in a patch release, so that
downstream users of `iroh@0.35` can make use of it.

## Breaking Changes
<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->
## Notes & open questions
<!-- Any notes, remarks or open questions you have to make about the PR.
-->
## Change checklist
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.