feat: add efficient metrics encoder that transfers schema only on change (#28)
## Description

Adds an `Encoder` that takes a registry and can emit `Update`s. An
`Update` is a serializable struct that has an optional `Schema` and a
required `Values`. `Schema` is only sent if the registry was changed
after the last send, and contains the name, type, help, prefixes, and
labels for all registered metrics. `Values` is just a vec of metric
values. So after the schema was sent in the first message, subsequent
metrics updates only contain the bare values, so are small.

Also adds a `Decoder` that can ingest the `Update`s and then allows
iterating over all metrics (basically zipping the current schema with
the current values).


## Breaking Changes
<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->
## Notes & open questions
<!-- Any notes, remarks or open questions you have to make about the PR.
-->
## Change checklist
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.